New implementation of grid based on OroDatagridBundle

+++ features/Context/FixturesContext.php
@@ -106,9 +106,12 @@ class FixturesContext extends RawMinkContext
      */
     public function clearPimFilesystem()
     {
+        // FIXME: Remove gitkeep?
         $fs = $this->getPimFilesystem();
         foreach ($fs->keys() as $key) {
-            $fs->delete($key);
+            if (strpos($key, '.') !== 0) {
+                $fs->delete($key);
+            }
         }
     }
 
@@ -247,6 +250,7 @@ class FixturesContext extends RawMinkContext
      * @param array|string $data
      *
      * @return Product
+     *
      * @Given /^a "([^"]*)" product$/
      */
     public function createProduct($data)
@@ -287,6 +291,7 @@ class FixturesContext extends RawMinkContext
      * @param string $sku
      *
      * @return Product
+     *
      * @Given /^(?:an|a) (enabled|disabled) "([^"]*)" product$/
      */
     public function anEnabledOrDisabledProduct($status, $sku)
@@ -856,6 +861,7 @@ class FixturesContext extends RawMinkContext
      * @param TableNode $table
      *
      * @return null
+     *
      * @Given /^the options "([^"]*)" of products (.*) should be:$/
      */
     public function theOptionsOfProductsShouldBe($attribute, $products, TableNode $table)
@@ -931,6 +937,7 @@ class FixturesContext extends RawMinkContext
      * @param TableNode $table
      *
      * @return null
+     *
      * @Given /^the following CSV to import:$/
      */
     public function theFollowingCSVToImport(TableNode $table)
@@ -1038,6 +1045,7 @@ class FixturesContext extends RawMinkContext
      * @param string $categoryCodes
      *
      * @return null
+     *
      * @Given /^(?:the )?categor(?:y|ies) of "([^"]*)" should be "([^"]*)"$/
      */
     public function theCategoriesOfShouldBe($productCode, $categoryCodes)
@@ -1114,6 +1122,7 @@ class FixturesContext extends RawMinkContext
      * @param string $username
      *
      * @return User
+     *
      * @Then /^there should be a "([^"]*)" user$/
      */
     public function getUser($username)
